backend/requirements.txt
Django
graphene-django
django-filter
django-graphql-jwt
django-cors-headers
psycopg2-binary
